+++ src/main/java/roomescape/controller/time/TimeController.java
@@ -10,6 +10,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
+import roomescape.controller.time.dto.AvailabilityTimeResponse;
+import roomescape.controller.time.dto.ReadTimeResponse;
+import roomescape.controller.time.dto.TimeRequest;
 import roomescape.service.TimeService;
 
 import java.net.URI;
