+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -6,10 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.controller.reservation.ReservationRequest;
-import roomescape.controller.reservation.ReservationResponse;
-import roomescape.controller.theme.ReservationThemeResponse;
-import roomescape.controller.time.AvailabilityTimeResponse;
+import roomescape.controller.reservation.dto.ReservationRequest;
+import roomescape.controller.reservation.dto.ReservationResponse;
+import roomescape.controller.theme.dto.ReservationThemeResponse;
+import roomescape.controller.time.dto.AvailabilityTimeResponse;
 import roomescape.service.exception.DuplicateReservation;
 import roomescape.service.exception.PreviousTimeException;
 import roomescape.service.exception.TimeNotFoundException;
