+++ src/test/java/roomescape/service/TimeServiceTest.java
@@ -6,9 +6,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.controller.time.AvailabilityTimeResponse;
-import roomescape.controller.time.ReadTimeResponse;
-import roomescape.controller.time.TimeRequest;
+import roomescape.controller.time.dto.AvailabilityTimeResponse;
+import roomescape.controller.time.dto.ReadTimeResponse;
+import roomescape.controller.time.dto.TimeRequest;
 import roomescape.repository.H2ReservationRepository;
 import roomescape.repository.H2ReservationTimeRepository;
 import roomescape.service.exception.TimeUsedException;
